Restructure internal wasm structures. Rename token class enum items to TC_... to differ from keywords ( T_... ).
[git-p4: depot-paths = "//depot/openwatcom/": change = 36475]

+++ bld/wasm/c/tbyte.c
@@ -311,7 +311,7 @@ static int multiply(ELD *op1, ELD *op2, ELD *res)
     res->m32[0] = r1.m32[3];
     res->m32[1] = r1.m32[4];
     res->m32[2] = r1.m32[5];
-    res->e = exp;
+    res->e = (unsigned short)exp;
     return 0;
 }
